# docs/web_deploy/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Blobs",
    "description": ""
  },
  "servers": [
    {
      "url": "https://api.demo.tokend.io",
      "description": "TokenD Developer Environment"
    }
  ],
  "paths": {
    "/blobs": {
      "post": {
        "tags": [
          "Blob"
        ],
        "summary": "Blobs",
        "description": "Create blob",
        "operationId": "createBlob",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": true,
                      "$ref": "#/components/schemas/Blobs"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Blob"
        ],
        "summary": "Blobs",
        "description": "Get blobs list",
        "operationId": "getBlobs",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": true,
                      "$ref": "#/components/schemas/Blobs"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/blobs/{id}": {
      "patch": {
        "tags": [
          "Blob"
        ],
        "operationId": "updateBlob",
        "summary": "Update Blob",
        "description": "Allows to update blob by id\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object",
                    "$ref": "#/components/schemas/Blobs"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad input parameter",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Blob"
        ],
        "operationId": "deleteBlob",
        "summary": "Delete Blob",
        "description": "Delete blob by id",
        "responses": {
          "204": {
            "description": "Blob deleted"
          },
          "400": {
            "description": "invalid blob id",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "500": {
            "description": "failed to get blob or failed to mark blob deleted",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlobKey": {
        "type": "object",
        "required": [
          "id",
          "type",
          "content"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "blobs"
            ]
          }
        }
      },
      "Blobs": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BlobKey"
          },
          {
            "type": "object",
            "required": [
              "attributes"
            ],
            "properties": {
              "attributes": {
                "type": "object",
                "required": [
                  "owner",
                  "content"
                ],
                "properties": {
                  "owner": {
                    "type": "string",
                    "description": "blob owner"
                  },
                  "content": {
                    "type": "string",
                    "description": "custom json"
                  }
                }
              }
            }
          }
        ]
      },
      "Errors": {
        "description": "Standard JSON:API error",
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "description": "Non empty array of errors occurred during request processing",
            "items": {
              "type": "object",
              "required": [
                "title",
                "status"
              ],
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Title is a short, human-readable summary of the problem",
                  "example": "Bad Request"
                },
                "detail": {
                  "type": "string",
                  "description": "Detail is a human-readable explanation specific to this occurrence of the problem",
                  "example": "Request body was expected"
                },
                "status": {
                  "type": "integer",
                  "description": "Status is the HTTP status code applicable to this problem",
                  "example": 400,
                  "enum": [
                    400,
                    401,
                    403,
                    404,
                    409,
                    500
                  ]
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "pageLimitParam": {
        "in": "query",
        "name": "page[limit]",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 15,
          "description": "Numbers of items per page to return."
        }
      },
      "pageNumberParam": {
        "in": "query",
        "name": "page[number]",
        "required": false,
        "schema": {
          "type": "integer"
        },
        "description": "The number of a page to return."
      },
      "sortingParam": {
        "in": "query",
        "name": "page[order]",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "desc",
          "description": "Order of records on the page. If sortingParam is not specified, order of records is by default sorted by ID."
        }
      }
    }
  }
}